requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
pymongo==4.5.0
gunicorn==21.2.0
python-dotenv==1.0.0